// clients/company-supervisor/src/types/InternshipEvalFromType.tsx
export type InternshipEvalFrom = {
  id: number;
  name: string;
  surname: string;
  question1: number;
  question2: number;
  question3: number;
  question4: number;
  question5: number;
  question6: number;
  question7: number;
  question8: number;
  internshipId: number;
  companyId: number;
};